+++ src/Opulence/Orm/DataMappers/ICachedSqlDataMapper.php
@@ -43,28 +43,6 @@ interface ICachedSqlDataMapper extends IDataMapper
     public function getSqlDataMapper(): SqlDataMapper;
 
     /**
-     * Gets a list of entities that differ in cache and the SQL database
-     *
-     * @return object[] The list of entities that were not already synced
-     *      The "missing" list contains the entities that were not in cache
-     *      The "differing" list contains the entities in cache that were not the same as SQL
-     *      The "additional" list contains entities in cache that were not at all in SQL
-     * @throws OrmException Thrown if there was an error getting the unsynced entities
-     */
-    public function getUnsyncedEntities(): array;
-
-    /**
-     * Refreshes the data in cache with the data from the SQL data mapper
-     *
-     * @return object[] The list of entities that were not already synced
-     *      The "missing" list contains the entities that were not in cache
-     *      The "differing" list contains the entities in cache that were not the same as SQL
-     *      The "additional" list contains entities in cache that were not at all in SQL
-     * @throws OrmException Thrown if there was an error refreshing the cache
-     */
-    public function refreshCache(): array;
-
-    /**
      * Refreshes an entity in cache with the entity from the SQL data mapper
      *
      * @param int|string $id The Id of the entity to sync